rename and move layout
